pyproject.toml
[project]
name = "emebeddingservice"
version = "0.1.0"
description = "Multi-modal embedding service for SlimRAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "markitdown>=0.1.2",
    "mineru[core]>=2.1.10",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "pillow>=9.0.0",
    "duckdb>=0.9.0",
    "meilisearch>=0.28.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]
